implemented a misc function to pypesto visualize to avoid code duplication
