+++ pypesto/visualize/clust_color.py
@@ -156,44 +156,3 @@ def assign_colors(vals, colors=None):
            'list of [r, g, b, alpha] with length equal to function '
            'values Number of function (here: ' + str(n_vals) + '), or as '
            'one single [r, g, b, alpha] color.')
-
-
-def handle_result_list(results, colors=None):
-    """
-    assigns colors to a list of results
-
-    Parameters
-    ----------
-
-    results: list or pypesto.Result
-        list of pypesto.Result objects or a single pypesto.Result
-
-    colors: list, optional
-        list of RGB colors
-
-    Returns
-    -------
-
-    results: pypesto.result or list
-       list of pypesto.result objects
-
-    colors: list of RGB
-        One for each element in 'results'.
-    """
-
-    # check if list
-    if not isinstance(results, list):
-        results = [results]
-    else:
-        # if more than one result is passed, one color per result is used
-        if colors is None:
-            colors = assign_colors(range(len(results)), colors)
-
-        # if the user passed a list of colors, does it have the correct length?
-        if any(isinstance(i_color, list) for i_color in colors):
-            if len(colors) != len(colors):
-                raise ('List of results and list of colors is passed. The '
-                       'length of the color list must match he length of the '
-                       'results list. Stopping.')
-
-    return results, colors
